sync with latest github master

+++ qCC/ccPointListPickingDlg.cpp
@@ -1,14 +1,14 @@
 //##########################################################################
 //#                                                                        #
-//#                            CLOUDCOMPARE                                #
+//#                              CLOUDCOMPARE                              #
 //#                                                                        #
 //#  This program is free software; you can redistribute it and/or modify  #
 //#  it under the terms of the GNU General Public License as published by  #
-//#  the Free Software Foundation; version 2 of the License.               #
+//#  the Free Software Foundation; version 2 or later of the License.      #
 //#                                                                        #
 //#  This program is distributed in the hope that it will be useful,       #
 //#  but WITHOUT ANY WARRANTY; without even the implied warranty of        #
-//#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         #
+//#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the          #
 //#  GNU General Public License for more details.                          #
 //#                                                                        #
 //#          COPYRIGHT: EDF R&D / TELECOM ParisTech (ENST-TSI)             #
@@ -30,7 +30,6 @@
 
 //qCC_db
 #include <ccLog.h>
-#include <ccObject.h> //for CC_QT5 def
 #include <ccPointCloud.h>
 #include <cc2DLabel.h>
 #include <ccPolyline.h>
@@ -60,7 +59,6 @@ ccPointListPickingDlg::ccPointListPickingDlg(QWidget* parent)
 	, m_orderedLabelsContainer(0)
 {
 	setupUi(this);
-	setWindowFlags(Qt::FramelessWindowHint | Qt::Tool);
 
 	exportToolButton->setPopupMode(QToolButton::MenuButtonPopup);
 	QMenu* menu = new QMenu(exportToolButton);
@@ -71,11 +69,7 @@ ccPointListPickingDlg::ccPointListPickingDlg(QWidget* parent)
 	QAction* exportToNewPolyline = menu->addAction("new polyline");
 	exportToolButton->setMenu(menu);
 
-#ifdef CC_QT5
 	tableWidget->verticalHeader()->setSectionResizeMode(QHeaderView::ResizeToContents);
-#else
-	tableWidget->verticalHeader()->setResizeMode(QHeaderView::ResizeToContents);
-#endif
 
 	startIndexSpinBox->setValue(s_pickedPointsStartIndex);
 	showGlobalCoordsCheckBox->setChecked(s_showGlobalCoordsCheckBoxChecked);
@@ -177,7 +171,10 @@ void ccPointListPickingDlg::cancelAndExit()
 	if (m_orderedLabelsContainer)
 	{
 		//Restore previous state
-		dbRoot->removeElements(m_toBeAdded);
+		if (!m_toBeAdded.empty())
+		{
+			dbRoot->removeElements(m_toBeAdded);
+		}
 
 		for (size_t j=0; j<m_toBeDeleted.size(); ++j)
 		{
@@ -289,7 +286,7 @@ void ccPointListPickingDlg::exportToNewPolyline()
 
 void ccPointListPickingDlg::applyAndExit()
 {
-	if (m_associatedCloud)
+	if (m_associatedCloud && !m_toBeDeleted.empty())
 	{
 		//apply modifications
 		MainWindow::TheInstance()->db()->removeElements(m_toBeDeleted); //no need to redraw as they should already be invisible
@@ -435,7 +432,7 @@ void ccPointListPickingDlg::exportToASCII(ExportFormat format)
 	}
 
 	//starting index
-	int startIndex = startIndexSpinBox->value();
+	unsigned startIndex = static_cast<unsigned>(std::max(0, startIndexSpinBox->value()));
 
 	for (unsigned i=0; i<count; ++i)
 	{
@@ -444,7 +441,7 @@ void ccPointListPickingDlg::exportToASCII(ExportFormat format)
 		const CCVector3* P = PP.cloud->getPoint(PP.index);
 
 		if (format == PLP_ASCII_EXPORT_IXYZ)
-			fprintf(fp,"%i,",i+startIndex);
+			fprintf(fp,"%u,",i+startIndex);
 		else if (format == PLP_ASCII_EXPORT_LXYZ)
 			fprintf(fp,"%s,",qPrintable(labels[i]->getName()));
 
@@ -512,16 +509,15 @@ void ccPointListPickingDlg::processPickedPoint(ccPointCloud* cloud, unsigned poi
 	newLabel->addPoint(cloud,pointIndex);
 	newLabel->setVisible(true);
 	newLabel->setDisplayedIn2D(false);
-	newLabel->setDisplayedIn3D(true);
+	newLabel->displayPointLegend(true);
 	newLabel->setCollapsed(true);
 	ccGenericGLDisplay* display = m_associatedCloud->getDisplay();
 	if (display)
 	{
 		newLabel->setDisplay(display);
-		int vp[4];
-		display->getViewportArray(vp);
-		newLabel->setPosition(	static_cast<float>(x+20)/static_cast<float>(vp[2]-vp[0]),
-								static_cast<float>(y+20)/static_cast<float>(vp[3]-vp[1]) );
+		QSize size = display->getScreenSize();
+		newLabel->setPosition(	static_cast<float>(x+20) / size.width(),
+								static_cast<float>(y+20) / size.height() );
 	}
 
 	//add default container if necessary
